fix(postgres): Clean up and add `tzdata` and `ca-certificates`
This commit fixes an issue where time zone data was not present
in the official Postgres image or latest CA certificates.  This
commit tidies up the `Dockerfile` to make its syntax more alike
with others in this repository and adds the aforementioned
packages.  This latter solves connection issues which may occur
from connecting to the Postgres instance from an IDE or
database manager tool.

Signed-off-by: Alexander Jung <alex@unikraft.cloud>